Extensions: fix two migration defects sqlite cannot catch
Both surfaced only on MySQL, because the test suite runs on sqlite and sqlite
enforces neither constraint.

The tombstone unique index generated a 68-character identifier and MySQL caps
them at 64, so `migrate` failed partway through creating the hook tables. Named
explicitly instead. I compiled all seven of this branch's migrations against the
MySQL grammar offline to check the rest; this was the only one over the limit.

extension_trusted_keys.repository_id was unsignedInteger while
extension_repositories.id is bigint unsigned, and MySQL refuses a foreign key
whose column type differs from the one it references. sqlite ignores the
mismatch entirely, so the constraint simply never existed there. The other two
new foreign keys point at users.id, which really is int unsigned, so they were
already correct.

Worth noting for future migrations on this branch: MySQL commits DDL
implicitly, so a migration that fails partway leaves its earlier tables behind
with no row in `migrations`, and the retry then fails on "table already exists"
rather than on the original defect.

+++ database/migrations/2026_09_08_000007_create_extension_signing_tables.php
@@ -29,7 +29,11 @@ return new class () extends Migration {
             $table->char('fingerprint', 64)->index();
             // Which repository presented this key, for the audit trail; trust
             // comes from the root signature, never from the source.
-            $table->unsignedInteger('repository_id')->nullable();
+            //
+            // bigInteger to match extension_repositories.id. MySQL refuses a
+            // foreign key whose column type differs from the one it references;
+            // sqlite does not, so the test suite cannot catch a mismatch here.
+            $table->unsignedBigInteger('repository_id')->nullable();
             $table->string('label')->nullable();
             $table->timestamp('valid_from')->nullable();
             $table->timestamp('valid_until')->nullable();
